ci: restore builder.Makefile for docker-ce-packaging
Compose gets rebuilt for `docker-ce-packaging` as part of DEB &
RPM builds. These cannot use Docker within the build, so need a
native `go build` command.

We want to keep this here so that we have control of the build
flags etc independently of packaging.

NOTE: The output path is now `bin/build/docker-compose` to keep
      parity with the rest of the Compose build logic - previously
      it was `bin/docker-compose` (no intermediate `build/` dir)

Signed-off-by: Milas Bowman <milas.bowman@docker.com>