watch: fix missing events from symlinked path (e.g. tmp on macOS)
Fix for watching for files in paths that have symlinks in them.

Resolve the project directory before joining it to the trigger
path. Additionally, resolve symlinks in any absolute trigger
paths.

There's probably (definitely) still some edge cases in here,
but this should fix the most common case, which is temporary
directories on macOS, which are symlinked to their real location
under `/private`.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>